+++ scripts/sync_h3lib.sh
@@ -11,3 +11,6 @@ cp -R h3_ffi/c/h3lib h3_flutter/android/src/h3lib
 
 rm -rf h3_flutter/windows/include/h3lib
 cp -R h3_ffi/c/h3lib h3_flutter/windows/include/h3lib
+
+rm -rf h3_flutter/linux/include/h3lib
+cp -R h3_ffi/c/h3lib h3_flutter/linux/include/h3lib